Added: Office365 connectivity
Added the ability to connect to Office365. This uses a JSON file as a configuration file.

At the moment, the Office 365 is a basic authentication where a username and password is passed, next step is oAuth.

Only exchange process so far is finding the Parent Folder, Awaiting, Failed, Successful Folders and receive the latest 100 emails from the Awaiting Folder.