Added: Outlook Folders
The way Arkes connects, you can specify a parent folder, this can be a folder called "Inbox" for the inbox, or another for a subfolder within Inbox.

Within that folder, this will be your process, and therefore should have an awaiting, failed, or successful. This is so we can grab emails from Awaiting, perform our actions, and move them to successful or failed.